Don't need to crosswalk these
Don't need to deal with  these schema because we will probably just be getting metatdata out in the DV elements JSON form and crosswalking ISO-19115 from that. If we need metadata from DV in these schema, DV can already export in these schema, so we would not need to do any crosswalking.